Exam Web/Bonus/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
PyJWT==2.8.0
Werkzeug==2.3.7
pyodbc==4.0.39
SQLAlchemy==2.0.21
python-dotenv==1.0.0 